feat(health): upgrade /health to components v2 UI + move to admin folder
